update func and functional buttons

+++ src/App03/pages/Places.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { FaTrash, FaEdit} from 'react-icons/fa'
+import { FaTrash, FaEdit, FaSearchLocation} from 'react-icons/fa'
+import {BiPhotoAlbum} from 'react-icons/bi'
 import {useState, useEffect, useRef} from 'react'
 import '../App03.css'
 
@@ -8,6 +9,7 @@ import {db} from '../firebaseConfig'
 import { doc,collection,getDocs,updateDoc,addDoc,deleteDoc} 
 from 'firebase/firestore'
 import { jsonEval } from '@firebase/util'
+import { HashRouter } from 'react-router-dom'
 
 
 function Places() {
@@ -17,6 +19,7 @@ const[pieces,setPieces] = useState([])
 const piecesCollectionRef= collection(db, 'pieces')
 //const [loading, setLoading] = useState(true)
 
+const [enableDel, setEnableDel] = useState(true)
 //Reffs
 //const  [input]
 const  inputName=useRef()
@@ -57,16 +60,45 @@ const createItm =async ()=>{
             {name:newName,place:newPlace,coors:newCoors,img:newImg,year:newYear,
             lat:newLat,lon:newLon})    }
 else{
-    console('edit')}
+
+    const editedItm={
+        name:inputName.current.value,
+         place:inputPlace.current.value,
+          img:inputImg.current.value,
+           year:inputYear.current.value,
+            lat: inputLat.current.value,
+             lon:inputLon.current.value}
+    //build doc obj to db
+    const itmDoc = doc(db,'pieces',currentId);
+    //updateDoc with editedItm    asynchronically (await)
+    await updateDoc(itmDoc, editedItm );
+    setEdit(false)
+    }
     window.location.reload(false)
 }
 
 const deleteItm= async (id)=>{
+   if (enableDel ) {
     const itmDoc= doc(db, 'pieces',id)
     await deleteDoc(itmDoc)
-    window.location.reload(false)
+    window.location.reload(false) }
+else{alert('editing has been disabled'  ) } 
 }
 
+
+const updateItm=(id,name,place,img,lat,lon,year,desc)=>{
+    setEdit(true)
+    // populate inputREFs with params of origin
+    inputName.current.value=name;
+    inputPlace.current.value=place;
+    inputImg.current.value=img;
+    inputYear.current.value=year;
+    inputLat.current.value=lat;
+    inputLon.current.value=lon;
+    
+    setCurrentId(id)
+
+}
 
   return (<>
     <div>Places</div>
@@ -111,7 +143,7 @@ onClick={createItm}>
         {pieces.map(  ({id,name,place,img,lat,lon,year,desc}) =>{
             return (<>
         <div className='cart'>
-                <p>{name}</p>
+                <h3 className='itm-title'>{name}</h3>
                 <p>{place}</p>
                 <p>{year}</p>
                 <img src={img} alt='pict'/>
@@ -119,11 +151,32 @@ onClick={createItm}>
                     <a href={`https://www.google.ca/maps/@${lat},${lon},250m/data=!3m1!1e3`} 
                      target="blank" >map it</a> 
                         </div>
-
-                <button          >{<FaEdit />}</button>
+<br></br>
+<hr></hr>
                 <button 
-                onClick={()=>{deleteItm(id)}  } > {<FaTrash />}  </button>
+                onClick={ ()=>{updateItm(id,name,place,img,lat,lon,year )}}
+                >{<FaEdit />}</button>
+                <button 
+                onClick=    {()=>{
+                    if (window.confirm('sure on deleting?') ) 
+                    deleteItm(id)}  }  > {<FaTrash />}  </button>
+             <a href={`https://www.google.ca/maps/@${lat},${lon},250m/data=!3m1!1e3`} 
+                     target="blank" >   <button >
+                {<    FaSearchLocation />}
+                </button>  </a>
+                <a href={`https://www.google.com/search?q=${name}+images&client=ubuntu&hs=doN&channel=fs&source=lnms&tbm=isch&sa=X&ved=2ahUKEwjC9IDJ_Z76AhWCuaQKHURyCtgQ_AUoAXoECAEQAw&biw=1209&bih=801&dpr=1`} 
+                     target="blank" >   <button >
+                {<BiPhotoAlbum/>}
+                </button>  </a>
+                <a href={`https://en.wikipedia.org/wiki/${name}`} 
+                     target="blank" >   <button >
+                {'W'}
+                </button>  </a>
 
+
+                
+
+                
                 </div>
             </>)
         } ) }
